Move class to own file
