+++ python/core/auto_generated/processing/models/qgsprocessingmodelresult.sip.in
@@ -1,0 +1,99 @@
+/************************************************************************
+ * This file has been generated automatically from                      *
+ *                                                                      *
+ * src/core/processing/models/qgsprocessingmodelresult.h                *
+ *                                                                      *
+ * Do not edit manually ! Edit header and run scripts/sipify.pl again   *
+ ************************************************************************/
+
+
+
+
+class QgsProcessingModelChildAlgorithmResult
+{
+%Docstring(signature="appended")
+Encapsulates the results of running a child algorithm within a model
+
+.. versionadded:: 3.38
+%End
+
+%TypeHeaderCode
+#include "qgsprocessingmodelresult.h"
+%End
+  public:
+
+    QgsProcessingModelChildAlgorithmResult();
+
+    Qgis::ProcessingModelChildAlgorithmExecutionStatus executionStatus() const;
+%Docstring
+Returns the status of executing the child algorithm.
+
+.. seealso:: :py:func:`setExecutionStatus`
+%End
+
+    void setExecutionStatus( Qgis::ProcessingModelChildAlgorithmExecutionStatus status );
+%Docstring
+Sets the ``status`` of executing the child algorithm.
+
+.. seealso:: :py:func:`executionStatus`
+%End
+
+    QVariantMap inputs() const;
+%Docstring
+Returns the inputs used for the child algorithm.
+
+.. seealso:: :py:func:`setInputs`
+%End
+
+    void setInputs( const QVariantMap &inputs );
+%Docstring
+Sets the ``inputs`` used for the child algorithm.
+
+.. seealso:: :py:func:`inputs`
+%End
+
+    QVariantMap outputs() const;
+%Docstring
+Returns the outputs generated by the child algorithm.
+
+.. seealso:: :py:func:`setOutputs`
+%End
+
+    void setOutputs( const QVariantMap &outputs );
+%Docstring
+Sets the ``outputs`` used for the child algorithm.
+
+.. seealso:: :py:func:`outputs`
+%End
+
+    QString htmlLog() const;
+%Docstring
+Returns the HTML formatted contents of logged messages which occurred while running the child.
+
+.. seealso:: :py:func:`setHtmlLog`
+%End
+
+    void setHtmlLog( const QString &log );
+%Docstring
+Sets the HTML formatted contents of logged messages which occurred while running the child.
+
+.. seealso:: :py:func:`htmlLog`
+%End
+
+    bool operator==( const QgsProcessingModelChildAlgorithmResult &other ) const;
+    bool operator!=( const QgsProcessingModelChildAlgorithmResult &other ) const;
+
+};
+
+
+
+
+
+
+/************************************************************************
+ * This file has been generated automatically from                      *
+ *                                                                      *
+ * src/core/processing/models/qgsprocessingmodelresult.h                *
+ *                                                                      *
+ * Do not edit manually ! Edit header and run scripts/sipify.pl again   *
+ ************************************************************************/
