+++ src/core/processing/models/qgsprocessingmodelresult.cpp
@@ -1,0 +1,25 @@
+/***************************************************************************
+                         qgsprocessingmodelresult.cpp
+                         ----------------------
+    begin                : April 2024
+    copyright            : (C) 2024 by Nyall Dawson
+    email                : nyall dot dawson at gmail dot com
+ ***************************************************************************/
+
+/***************************************************************************
+ *                                                                         *
+ *   This program is free software; you can redistribute it and/or modify  *
+ *   it under the terms of the GNU General Public License as published by  *
+ *   the Free Software Foundation; either version 2 of the License, or     *
+ *   (at your option) any later version.                                   *
+ *                                                                         *
+ ***************************************************************************/
+
+#include "qgsprocessingmodelresult.h"
+
+//
+// QgsProcessingModelChildResult
+//
+
+QgsProcessingModelChildAlgorithmResult::QgsProcessingModelChildAlgorithmResult() = default;
+
